+++ QuodLib.IO/IOTasks.cs
@@ -5,8 +5,10 @@ using System.Text;
 using System.Threading.Tasks;
 
 using System.IO;
+using QuodLib.IO.Symbolic;
 
-namespace QuodLib.IO {
+namespace QuodLib.IO
+{
     /// <summary>
     /// A class of methods for running System.IO methods asynchronously.
     /// </summary>
@@ -19,7 +21,7 @@ namespace QuodLib.IO {
             => RunWithRethrow(() => Directory.Exists(path));
 
         public static Task<bool> Path_IsSymbolicAsync(string path)
-            => RunWithRethrow(() => Symbolic.IsSymbolic(path));
+            => RunWithRethrow(() => Info.IsSymbolic(path));
 
         public static Task<string[]> GetDirectoriesAsync(string path)
             => RunWithRethrow(() => Directory.GetDirectories(path));
